# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnHub learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/checkout:
    post:
      operationId: createCheckoutSession
      tags: [subscriptions, esewa]
      summary: Create an eSewa checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /subscriptions/verify:
    post:
      operationId: verifyPayment
      tags: [subscriptions, esewa]
      summary: Verify eSewa payment and activate subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EsewaVerifyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsewaVerifyResult"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CheckoutSessionInput:
      type: object
      required: [plan]
      properties:
        plan: { type: string }
    CheckoutSession:
      type: object
      required: [paymentUrl, formData]
      properties:
        paymentUrl: { type: string }
        formData:
          type: object
          required: [amount, tax_amount, total_amount, transaction_uuid, product_code, product_service_charge, product_delivery_charge, success_url, failure_url, signed_field_names, signature]
          properties:
            amount: { type: string }
            tax_amount: { type: string }
            total_amount: { type: string }
            transaction_uuid: { type: string }
            product_code: { type: string }
            product_service_charge: { type: string }
            product_delivery_charge: { type: string }
            success_url: { type: string }
            failure_url: { type: string }
            signed_field_names: { type: string }
            signature: { type: string }
    EsewaVerifyInput:
      type: object
      required: [encodedData, plan]
      properties:
        encodedData: { type: string }
        plan: { type: string }
    EsewaVerifyResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }

    # ─── Dashboard ────────────────────────────────────────────────────────────
...